Agrega decimal al precio en la tabla de productos

+++ database/migrations/2021_10_10_192458_create_productos_table.php
@@ -19,7 +19,7 @@ class CreateProductosTable extends Migration
             $table->string('nombre', 100);
             $table->string('marca', 100);
             $table->string('deporte', 100);
-            $table->integer('precio');
+            $table->decimal('precio', 7, 2);
             $table->string('imagen', 100);
             $table->string('descripcionimg', 100);
         });
